feat: Add getUserChannelProfile endpoint
This commit adds a new endpoint `getUserChannelProfile` to the `user.controllers.js` file. This endpoint retrieves the channel profile information for a given username. It performs an aggregation query on the `User` collection to fetch the channel details, including the number of subscribers, channels subscribed to, and subscription status. If the channel does wr