feat: Add getWatchHistory endpoint
The `getWatchHistory` endpoint is added to the `user.controllers.js` file. This endpoint retrieves the watch history for a user by performing an aggregation query on the `User` collection. It fetches the user's watch history, including the video details and the owner's information. This feature allows users to view their previously watched videos.